Issue #2987418 by quietone, Kristen Pol: Rename MigrateUpgrade tests
(cherry picked from commit 898083ce7bd89e65360bb581ec174ee01c1b1e7f)
