+++ modules/migrate_drupal_ui/tests/src/Functional/d7/MultilingualReviewPageTest.php
@@ -2,15 +2,17 @@
 
 namespace Drupal\Tests\migrate_drupal_ui\Functional\d7;
 
-use Drupal\Tests\migrate_drupal_ui\Functional\MigrateUpgradeReviewPageTestBase;
+use Drupal\Tests\migrate_drupal_ui\Functional\MultilingualReviewPageTestBase;
 
 /**
  * Tests migrate upgrade review page for Drupal 7.
  *
+ * Tests with translation modules and migrate_drupal_multilingual enabled.
+ *
  * @group migrate_drupal_7
  * @group migrate_drupal_ui
  */
-class MigrateUpgrade7ReviewPageTest extends MigrateUpgradeReviewPageTestBase {
+class MultilingualReviewPageTest extends MultilingualReviewPageTestBase {
 
   /**
    * {@inheritdoc}
