+++ modules/migrate_drupal_ui/tests/src/Functional/d7/Upgrade7Test.php
@@ -13,7 +13,7 @@ use Drupal\user\Entity\User;
  *
  * @group migrate_drupal_ui
  */
-class MigrateUpgrade7Test extends MigrateUpgradeExecuteTestBase {
+class Upgrade7Test extends MigrateUpgradeExecuteTestBase {
 
   /**
    * Modules to enable.
